Issue/129/refactoring cli (#130)
* change folder of the command line

* fix test

* black refactoring and pep8

* huge refactoring of the command line
