+++ fink_fat/associations/associations.py
@@ -424,7 +424,6 @@ def night_to_night_trajectory_associations(
     )
 
     if len(traj_assoc) != 0:
-
         traj_assoc, new_obs_assoc = cone_search_association(
             two_last_observations, traj_assoc, new_obs_assoc, angle_criterion
         )
@@ -525,7 +524,6 @@ def tracklets_and_trajectories_associations(
     if len(trajectories) == 0 or len(tracklets) == 0:
         return trajectories, tracklets, max_traj_id
     else:
-
         trajectories_not_updated = trajectories[trajectories["not_updated"]]
 
         # get the last two observations for each trajectories
@@ -549,7 +547,6 @@ def tracklets_and_trajectories_associations(
 
         # for each trajectory nid from the last observations
         for tr_nid in trajectories_nid:
-
             # get the two last observations with the tracklets extremity that have the current nid
             two_last_current_nid = two_last_observation_trajectory[
                 two_last_observation_trajectory["trajectory_id"].isin(
@@ -579,7 +576,6 @@ def tracklets_and_trajectories_associations(
             )
 
             if len(traj_extremity_associated) > 0:
-
                 # creates a dataframe for each duplicated trajectory associated with the tracklets
                 duplicates = traj_extremity_associated["trajectory_id"].duplicated()
 
@@ -587,7 +583,6 @@ def tracklets_and_trajectories_associations(
                 tracklets_duplicated = traj_extremity_associated[duplicates]
 
                 if len(tracklets_duplicated) > 0:
-
                     # get the trajectories involved in the duplicates
                     duplicate_traj = (
                         trajectories[
@@ -701,7 +696,6 @@ def tracklets_and_trajectories_associations(
 
                 # If too slow, need to be optimized/vectorized
                 for _, rows in traj_extremity_associated.iterrows():
-
                     # get all rows of the associated tracklets of the next night
                     next_night_tracklets = tracklets[
                         tracklets["trajectory_id"] == rows["tmp_traj"]
@@ -861,7 +855,6 @@ def trajectories_with_new_observations_associations(
     if len(trajectories) == 0 or len(new_observations) == 0:
         return trajectories, new_observations, max_traj_id
     else:
-
         # get only the trajectories not updated by a previous step
         trajectories_not_updated = trajectories[trajectories["not_updated"]]
 
@@ -881,7 +874,6 @@ def trajectories_with_new_observations_associations(
 
         # for each trajectory nid from the last observations
         for tr_nid in trajectories_nid:
-
             # get the two last observations with the tracklets extremity that have the current nid
             two_last_current_nid = two_last_observation_trajectory[
                 two_last_observation_trajectory["trajectory_id"].isin(
@@ -916,7 +908,6 @@ def trajectories_with_new_observations_associations(
             ]
 
             if len(obs_assoc) > 0:
-
                 # creates a dataframe for each duplicated trajectory associated with the tracklets
                 duplicates = obs_assoc["trajectory_id"].duplicated()
 
@@ -924,7 +915,6 @@ def trajectories_with_new_observations_associations(
                 duplicate_obs = obs_assoc[duplicates]
 
                 if len(duplicate_obs) > 0:
-
                     # get the trajectories involved in the duplicates
                     duplicate_traj = (
                         trajectories[
@@ -1101,7 +1091,6 @@ def old_observations_with_tracklets_associations(
     if len(tracklets) == 0 or len(old_observations) == 0:
         return tracklets, old_observations, max_traj_id
     else:
-
         # get all the old night id sort by descending order to begin the associations
         # with the recently ones
         old_obs_nid = np.sort(np.unique(old_observations["nid"]))[::-1]
@@ -1111,7 +1100,6 @@ def old_observations_with_tracklets_associations(
         )
 
         for obs_nid in old_obs_nid:
-
             current_old_obs = old_observations[old_observations["nid"] == obs_nid]
 
             diff_night = next_nid - obs_nid
@@ -1143,7 +1131,6 @@ def old_observations_with_tracklets_associations(
             ]
 
             if len(track_left_assoc) > 0:
-
                 # creates a dataframe for each duplicated trajectory associated with the tracklets
                 duplicates = old_obs_right_assoc["trajectory_id"].duplicated()
 
@@ -1151,7 +1138,6 @@ def old_observations_with_tracklets_associations(
                 duplicate_obs = old_obs_right_assoc[duplicates]
 
                 if len(duplicate_obs) > 0:
-
                     # get the tracklets involved with duplicates
                     duplicate_traj = (
                         tracklets[
@@ -1313,12 +1299,10 @@ def old_with_new_observations_associations(
     if len(old_observations) == 0 or len(new_observations) == 0:
         return pd.DataFrame(), old_observations, new_observations
     else:
-
         old_obs_nid = np.sort(np.unique(old_observations["nid"]))[::-1]
         trajectory_df = pd.DataFrame()
 
         for obs_nid in old_obs_nid:
-
             current_old_obs = old_observations[old_observations["nid"] == obs_nid]
 
             diff_night = next_nid - obs_nid
@@ -1336,7 +1320,6 @@ def old_with_new_observations_associations(
             )
 
             if len(left_assoc) > 0:
-
                 new_trajectory_id = np.arange(
                     last_trajectory_id, last_trajectory_id + len(left_assoc)
                 )
@@ -1504,7 +1487,6 @@ def time_window_management(
     oldest_traj = pd.DataFrame(columns=trajectory_df.columns)
 
     if len(trajectory_df) > 0:
-
         # get the lost observation of all trajectories
         last_obs_of_all_traj = get_n_last_observations_from_trajectories(
             trajectory_df, 1
@@ -1512,7 +1494,6 @@ def time_window_management(
 
         # if the night difference exceed the time window
         if (nid_next_night - last_nid > traj_time_window) and keep_last:
-
             # get last observation of trajectories from the last nid
             last_obs_of_all_traj = last_obs_of_all_traj[
                 last_obs_of_all_traj["nid"] == last_nid
@@ -1586,7 +1567,6 @@ def time_window_management(
 
 
 if __name__ == "__main__":  # pragma: no cover
-
     if "unittest.util" in __import__("sys").modules:
         # Show full diff in self.assertEqual.
         __import__("sys").modules["unittest.util"]._MAX_LENGTH = 999999999
