+++ fink_fat/command_line/offline_cli.py
