+++ fink_fat/command_line/stat_cli.py
